Move request parsing to the profiling process
TPS was impacted heavily by doing request parsing in the main http
server. This change moves that work to the profiling server and replaces
the pydantic json parsing with orjson, which just parses to a dict much
faster. This change made throughput go from 700tps -> 2000tps on an
m5.large.