Refactor dockerfile, add fail mode when missing config
Updated docker files to ignore most of the bulky local content when
copying build context and moved the src copy to a later layer so that
source updates don't trigger apt-get and poetry layers anymore.

Also add a new env var to fail the container startup if the config is
missing as a safeguard to image building mistakes.