Add embedding specific api,  upgrade to  whylogs multi dataset logger, request based password support (#4)
* Add request based password support

This enables password enforcement for each request. Include a header as
follows:

Authorization: Bearer my-password

* Upgrade to the newer whylogs multi dataset logger

This logger will do the work of managing multiple dataset
profiles/segments for different dataset timestamps. It makes the
timestamp field in the request actually do something.

* Add embedding specific api

The performance of embeddings isn't good enough to log it the way
everything else is logged right now. It requires special preprocessing
into a numpy matrix unfortunately. Hopefully we can get rid of that api
after we figure out the perf issue in whylogs.

* Refactoring, remove single/multiple field for single data field

Making it backwards compatible with the previous container adds a fair
bit of preprocessing to the batch message handling for no particularly
good reason. I'm just getting rid of the single/multiple option in favor
of a single data field that looks like multiple did.

---------

Co-authored-by: Anthony Naddeo <anthony@whylabs.ai>